:hammer: Rename unique.jsx to unique-minichart.jsx
