+++ src/internal-packages/schema/lib/component/minichart.jsx
@@ -1,7 +1,7 @@
 const app = require('hadron-app');
 const React = require('react');
 const PropTypes = require('prop-types');
-const UniqueMiniChart = require('./unique');
+const UniqueMiniChart = require('./unique-minichart');
 const _ = require('lodash');
 const DocumentMinichart = require('./document');
 const ArrayMinichart = require('./array');
