+++ src/internal-packages/schema/lib/component/unique-minichart.jsx
